made a display for failed files that includes being able to check them (so they can later be submitted as attachments to the collection)
